+++ src/main/java/com/ardaslegends/albaseplugin/tabcompletion/TabCompletionALReload.java
@@ -8,7 +8,7 @@ import org.bukkit.command.TabCompleter;
 import java.util.ArrayList;
 import java.util.List;
 
-public class TabCompletionReload  implements TabCompleter {
+public class TabCompletionALReload implements TabCompleter {
     @Override
     public List<String> onTabComplete(CommandSender sender, Command command, String label, String[] args) {